refactored models + notebook dir
