+++ restorers/model/mirnetv2/__init__.py
@@ -1,0 +1,1 @@
+from .mirnet import MirNetv2
